feat/refactor: CourseRepository and Repository.
CourseRepository was implemented and moved into Repositories folder as well as Repository.
